backend/api/requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-migrate>=4.0
flask-limiter>=3.5
flask-wtf>=1.2
flask-cors>=4.0
pillow>=10.0
werkzeug>=3.0
scikit-learn>=1.3
numpy>=1.24
python-dotenv>=1.0
gunicorn>=21.0
